# Avoid boxing when calling withValue, withValueOpt on IntEnum, ShortEnum