--- a/enumeratum-core/compat/src/main/scala-2.11/enumeratum/values/ValueEnum.scala
+++ b/enumeratum-core/compat/src/main/scala-2.11/enumeratum/values/ValueEnum.scala
@@ -1,15 +1,15 @@
 package enumeratum.values
 
 import enumeratum.ValueEnumMacros
-
 import scala.language.experimental.macros
 
-sealed trait ValueEnum[ValueType <: AnyVal, EntryType <: ValueEnumEntry[ValueType]] {
+sealed trait ValueEnum[@specialized(Short, Int) ValueType <: AnyVal, EntryType <: ValueEnumEntry[ValueType]] {
 
   /**
    * Map of [[ValueType]] to [[EntryType]] members
    */
   final lazy val valuesToEntriesMap: Map[ValueType, EntryType] = values.map(v => v.value -> v).toMap
+  private lazy val existingEntriesString = values.map(_.value).mkString(", ")
 
   /**
    * The sequence of values for your [[Enum]]. You will typically want
@@ -35,66 +35,8 @@ sealed trait ValueEnum[ValueType <: AnyVal, EntryType <: ValueEnumEntry[ValueTyp
    */
   def withValueOpt(i: ValueType): Option[EntryType] = valuesToEntriesMap.get(i)
 
-  private lazy val existingEntriesString = values.map(_.value).mkString(", ")
-
   protected def buildNotFoundMessage(i: ValueType): String = {
     s"$i is not a member of ValueEnum ($existingEntriesString)"
-  }
-
-}
-
-/**
- * An [[ValueEnum]] that has an optimized array access that does not uses
- * boxing when accessing the methods [[ValueEnum.withValue()]] and [[ValueEnum.withValueOpt()]]
- */
-trait IndexedEnum[@specialized(Int, Short) ValueType <: AnyVal, EntryType <: ValueEnumEntry[ValueType]]
-    extends ValueEnum[ValueType, EntryType] {
-
-  /**
-   * Converts the given value to [[Int]]
-   *
-   * @param value The value to convert
-   * @return The converted value
-   */
-  def toIndex(value: ValueType): Int
-
-  protected lazy val (arrayValues, minValue) = {
-    val indices = values.map(v => toIndex(v.value))
-
-    val minValue = indices.min
-    val maxValue = indices.max
-
-    val retVal = Array.fill[Option[EntryType]](maxValue - minValue + 1)(None)
-
-    for (value <- values) {
-      retVal(toIndex(value.value) - minValue) = Some(value)
-    }
-
-    retVal -> minValue
-  }
-
-  /**
-   * Optionally returns an [[Int]] for a given value.
-   */
-  override def withValueOpt(i: ValueType) = {
-    val index = toIndex(i) - minValue
-    if (index >= arrayValues.length || index < 0) None else arrayValues(index)
-  }
-
-  /**
-   * Tries to get an [[Int]] by the supplied value. The value corresponds to the .value
-   * of the case objects implementing [[Int]]
-   *
-   * Like [[Enumeration]]'s `withValue`, this method will throw if the value does not match any of the values'
-   * `.value` values.
-   */
-  override def withValue(i: ValueType): EntryType = {
-    val index = toIndex(i) - minValue
-    if (index >= arrayValues.length || index < 0) {
-      throw new NoSuchElementException(buildNotFoundMessage(i))
-    } else {
-      arrayValues(index).getOrElse(throw new NoSuchElementException(buildNotFoundMessage(i)))
-    }
   }
 
 }
@@ -110,7 +52,7 @@ trait IndexedEnum[@specialized(Int, Short) ValueType <: AnyVal, EntryType <: Val
 /**
  * Value enum with [[IntEnumEntry]] entries
  */
-trait IntEnum[A <: IntEnumEntry] extends IndexedEnum[Int, A] {
+trait IntEnum[A <: IntEnumEntry] extends ValueEnum[Int, A] {
 
   /**
    * Method that returns a Seq of [[A]] objects that the macro was able to find.
@@ -119,14 +61,6 @@ trait IntEnum[A <: IntEnumEntry] extends IndexedEnum[Int, A] {
    * if you aren't using this method...why are you even bothering with this lib?
    */
   protected def findValues: IndexedSeq[A] = macro ValueEnumMacros.findIntValueEntriesImpl[A]
-
-  /**
-   * Converts the given value to [[Int]]
-   *
-   * @param value The value to convert
-   * @return The converted value
-   */
-  override def toIndex(value: Int): Int = value
 }
 
 /**
@@ -147,7 +81,7 @@ trait LongEnum[A <: LongEnumEntry] extends ValueEnum[Long, A] {
 /**
  * Value enum with [[ShortEnumEntry]] entries
  */
-trait ShortEnum[A <: ShortEnumEntry] extends IndexedEnum[Short, A] {
+trait ShortEnum[A <: ShortEnumEntry] extends ValueEnum[Short, A] {
 
   /**
    * Method that returns a Seq of [[A]] objects that the macro was able to find.
@@ -156,13 +90,5 @@ trait ShortEnum[A <: ShortEnumEntry] extends IndexedEnum[Short, A] {
    * if you aren't using this method...why are you even bothering with this lib?
    */
   final protected def findValues: IndexedSeq[A] = macro ValueEnumMacros.findShortValueEntriesImpl[A]
-
-  /**
-   * Converts the given value to [[Int]]
-   *
-   * @param value The value to convert
-   * @return The converted value
-   */
-  override def toIndex(value: Short): Int = value.toInt
 
 }