Hi,

First of all, I didn't managed to run the tests successfully on the master branch, I hope that it relates to Scala.js. 

I've added a trait called `IndexedEnum`, This trait implements `withValue`, `withValueOpt` using the apply method of the indexedValues, Since the values are indexed, This trait is specialized for short and int, so calling the methods `withValue`, `withValueOpt` wouldn't cause boxing.

~~There is one thing, I'm not sure how the macro `ValueEnumMacros.findIntValueEntriesImpl[A]` works, And what happens if there is a gap between Enum Values (e.g, first value is 5).~~

I'm not sure if you want to make `ValueEnum` specialized for int and short, So there won't be boxing when people pass the `ValueEnum` as `ValueEnum` (and not as the real type)

The `ValueEnumSpec` passed and validated the changes :+1: