@shanielh 2016-06-29
Not sure why coverage is decreased, I haven't touched any of the tests :open_mouth: