@lloydmeta 2016-06-29
Hi there,

Thanks for your interest in this project. For my reference, was this PR made out of concern for performance? It helps for me to know your motivation :)

> First of all, I didn't managed to run the tests successfully on the master branch, I hope that it relates to Scala.js.

Really? Can you describe what errors you had? The tests run fine on my computer and on Travis.

The PR looks reasonable and I just have a few concerns:

- Looks like when values are negative, the PR as-is would not work because the values are used as array indices
- If we have enumerations where the range is quite big, we will end up with really big arrays filled with references to None
- No tests for added lines code. This is why the coverage percentage decreased.
- Might be nice to have a few benchmarks to show this complexity is worthwhile

A somewhat contrived example of something that might not work:

```scala
sealed abstract class Stuff(val value: Int) extends IntEnumEntry
object Stuff extends IntEnum[Stuff] {
  val values = findValues
  case object Beans(-1000) extends Stuff
  case object Other(Int.MaxValue) extends Stuff
}
```