@shanielh 2016-06-29
Hi @lloydmeta .

Yeah, The main concern was performance.

For your concerns : 

* We can add it as opt-in behavior, and not "mix in" the `IndexedEnum` to `IntEnum` and `ShortEnum` by default.
* Same as above
* I haven't added that line of code, It was there in the first place :confused: 
* I can add a test using [ScalaMeter](http://scalameter.github.io/) if you want.
```

[Log for sbt test](https://github.com/lloydmeta/enumeratum/files/338799/log.txt).

No problem, I'll add commits and squash them (if and) when you want to merge :)