@lloydmeta 2016-06-29
Cool I hear you on the performance concern. It might be worthwhile to add some performance tests then to show the difference in performance. If we add some benchmarks in this PR, then I would prefer we use JMH because it is battle tested by professionals ([see this thread for how easily things can be implemented wrong](https://groups.google.com/forum/#!topic/mechanical-sympathy/m4opvy4xq3U)).

Even though we can make this an opt-in behaviour for the sake of performance, it might be better to try to find a way to avoid non-intuitive behaviour that throws errors at runtime and is hard to test for in the first place. Is there some other way to achieve specialisation ?

BTW, this PR adds 72 lines of code.  For instance, see the green blocks [here](https://github.com/lloydmeta/enumeratum/pull/42/files#diff-b06bfbd455a9b7b83b17066ea3664b9bR46)