@shanielh 2016-06-29
Can you check in JMH the number of boxing in a specific code section? (I know we're able to do that in ScalaMeter).

We can add the "lower bound" of the array as a field and fill the array from the lower bound, That way, It would work with negative values.