@lloydmeta 2016-06-29
> Can you check in JMH the number of boxing in a specific code section? (I know we're able to do that in ScalaMeter).

I'm not sure about checking the number of boxing in a specific code section in JMH. It's pretty interesting that you can do that in Scalatest !

> We can add the "lower bound" of the array as a field and fill the array from the lower bound, That way, It would work with negative values.

That sounds like a reasonable approach, maybe we can go with that for now? Alternately we can maybe import the code from [Debox's Map](https://github.com/non/debox/blob/master/src/main/scala/debox/Map.scala).

> By looking at the coverage report, it says that number of covered lines in Cicle decreased - https://coveralls.io/builds/6795267

This is odd considering [this part of CirceSpec](https://github.com/lloydmeta/enumeratum/blob/master/enumeratum-circe/src/test/scala/enumeratum/CirceSpec.scala#L33-L35) should invoke that line...