@shanielh 2016-06-29
@lloydmeta So, work items for this PR : 

- [ ] me - Add test for negative sequences and fix it
- [ ] @lloydmeta - Read about Scalatest and see if it's good enough for you