@lloydmeta 2016-06-29
Scalameter sounds fine to me :)