@cb372 2016-06-29
If you want to avoid boxing, have you considered simply replacing `Map` with `LongMap`?