@shanielh 2016-06-29
@cb372 - Since that won't help with the boxing / unboxing, The `Trait` map is still of `[A, B]` and both of them are not specialized. `LongMap` may be faster than normal (hash/trie) `Map` when the key is Long, but boxing/unboxing will still happen.

@lloydmeta - I've rewrote the code, the coverage should be smaller now, I'll add tests for these, but before that, The reason I changed the way of implementation is : 

1. When deriving a trait from trait, `@specialized` isn't working good : http://axel22.github.io/2013/11/03/specialization-quirks.html
2. Now the implementation is "opt in", And if you have a long enum, You can use the new `IndexedEnumMap` too.

Actually, I can take this code and use it in my case, But it would be nice to merge this PR / even add documentation about `withValue` / `withValueOpt` and detail about their boxing and how to avoid it.

Tell me what you think about it. I can add benchmark to see that the newer way is faster than the old.