@cb372 2016-06-29
Sorry for my very terse comment, I was on my phone. I meant something like this. (Untested, probably doesn't compile.)

```scala
import scala.collection.immutable.LongMap

// Type class for things that can be treated as a Long
trait ToLong[A] { def toLong(value: A): Long }
object ToLong {
  implicit val longToLong  = new ToLong[Long] { def toLong(value: Long) = value }
  implicit val intToLong   = new ToLong[Int] { def toLong(value: Int) = value.toLong }
  implicit val shortToLong = new ToLong[Short] { def toLong(value: Short) = value.toLong }
}

sealed trait ValueEnum[ValueType : ToLong, EntryType <: ValueEnumEntry[ValueType]] {

  private val tl = implicitly[ToLong[ValueType]]

  final lazy val valuesToEntriesMap: LongMap[EntryType] = {
    val elems = values.map(v => tl.toLong(v.value) -> v)
    LongMap(elems)
  }

  def withValueOpt(i: ValueType): Option[EntryType] = valuesToEntriesMap.get(tl.toLong(i))
  
  // other methods unchanged

}
```

Benefits:

* No need to use `@specialized`, which is a tricky beast
* No risk of allocating an enormous sparse array

Performance impact:

* Lookups will now include the (tiny) overhead of a `i2l` instruction, but HotSpot can probably optimise it away.