@shanielh 2016-06-29
@cb372 Lookups will still include overhead of Boxing 😞 

I guess we just need to compare the performance of both of the ways to see what's faster. I hope I'll have time to sit on this on the weekend.