@cb372 2016-06-29
Oh yeah, `def toLong(value: A): Long` would box unless you `@specialized` it. Bleurgh.