@lloydmeta 2016-07-17
Sorry for the delayed response. I've added benchmarks to this project (JMH) and got the following:

## Before

```
[info] Benchmark                                            Mode  Cnt     Score    Error  Units
[info] values.ValueEnumBenchmarks.withValueDoesNotExist     avgt   30  1707.968 ± 31.495  ns/op
[info] values.ValueEnumBenchmarks.withValueExists           avgt   30     3.941 ±  0.033  ns/op
[info] values.ValueEnumBenchmarks.withValueOptDoesNotExist  avgt   30     5.435 ±  0.033  ns/op
[info] values.ValueEnumBenchmarks.withValueOptExists        avgt   30     6.313 ±  0.032  ns/op
```

## After 

```
[info] Benchmark                                            Mode  Cnt     Score    Error  Units
[info] values.ValueEnumBenchmarks.withValueDoesNotExist     avgt   30  1679.596 ± 78.947  ns/op
[info] values.ValueEnumBenchmarks.withValueExists           avgt   30     9.227 ±  0.065  ns/op
[info] values.ValueEnumBenchmarks.withValueOptDoesNotExist  avgt   30     5.554 ±  0.039  ns/op
[info] values.ValueEnumBenchmarks.withValueOptExists        avgt   30     6.453 ±  0.037  ns/op
```

It's entirely possible that my benchmarks are flawed (see [here](https://github.com/lloydmeta/enumeratum/blob/master/benchmarking/compat/src/main/scala-2.11/enumeratum/values/ValueEnumBenchmarks.scala)), but it doesn't look like the changes here have improved performance in `ValueEnum`s.